// main/src/types/problem.ts
export interface Problem {
  id: string;
  name: string;
  content: string;
  input: string[];
  output: string[];
  created_at: string;
  updated_at: string;
  is_active?: boolean; // optional until ensured in all selects
  time_limit?: number; // Time limit in milliseconds
  memory_limit?: number; // Memory limit in MB
  points: number;
}

export interface Contest {
  id: string;
  name: string;
  description: string | null;
  length: number;
  created_at: string;
  updated_at: string;
}